+++ internal/encoding/encoding.go
@@ -30,6 +30,7 @@ var uvarintPool = &sync.Pool{
 
 // decodeBytes decodes a varint length-prefixed byte slice, returning it along with the number
 // of input bytes read.
+// Assumes bz will not be mutated.
 func DecodeBytes(bz []byte) ([]byte, int, error) {
 	s, n, err := DecodeUvarint(bz)
 	if err != nil {
@@ -51,9 +52,7 @@ func DecodeBytes(bz []byte) ([]byte, int, error) {
 	if len(bz) < end {
 		return nil, n, fmt.Errorf("insufficient bytes decoding []byte of length %v", size)
 	}
-	bz2 := make([]byte, size)
-	copy(bz2, bz[n:end])
-	return bz2, end, nil
+	return bz[n:end], end, nil
 }
 
 // decodeUvarint decodes a varint-encoded unsigned integer from a byte slice, returning it and the
@@ -90,6 +89,23 @@ func DecodeVarint(bz []byte) (int64, int, error) {
 // EncodeBytes writes a varint length-prefixed byte slice to the writer.
 func EncodeBytes(w io.Writer, bz []byte) error {
 	err := EncodeUvarint(w, uint64(len(bz)))
+	if err != nil {
+		return err
+	}
+	_, err = w.Write(bz)
+	return err
+}
+
+var hashLenBz []byte
+
+func init() {
+	hashLenBz = make([]byte, 1)
+	binary.PutUvarint(hashLenBz, 32)
+}
+
+// Encode 32 byte long hash
+func Encode32BytesHash(w io.Writer, bz []byte) error {
+	_, err := w.Write(hashLenBz)
 	if err != nil {
 		return err
 	}
@@ -139,6 +155,10 @@ func EncodeUvarintSize(u uint64) int {
 
 // EncodeVarint writes a varint-encoded integer to an io.Writer.
 func EncodeVarint(w io.Writer, i int64) error {
+	if bw, ok := w.(io.ByteWriter); ok {
+		return fVarintEncode(bw, i)
+	}
+
 	// Use a pool here to reduce allocations.
 	//
 	// Though this allocates just 10 bytes on the stack, doing allocation for every calls
@@ -155,6 +175,26 @@ func EncodeVarint(w io.Writer, i int64) error {
 	varintPool.Put(buf)
 
 	return err
+}
+
+func fVarintEncode(bw io.ByteWriter, x int64) error {
+	// Firstly convert it into a uvarint
+	ux := uint64(x) << 1
+	if x < 0 {
+		ux = ^ux
+	}
+	for ux >= 0x80 { // While there are 7 or more bits in the value, keep going
+		// Convert it into a byte then toggle the
+		// 7th bit to indicate that more bytes coming.
+		// byte(x & 0x7f) is redundant but useful for illustrative
+		// purposes when translating to other languages
+		if err := bw.WriteByte(byte(ux&0x7f) | 0x80); err != nil {
+			return err
+		}
+		ux >>= 7
+	}
+
+	return bw.WriteByte(byte(ux & 0x7f))
 }
 
 // EncodeVarintSize returns the byte size of the given integer as a varint.