+++ node.go
@@ -57,16 +57,21 @@ func GetRootKey(version int64) []byte {
 
 // Node represents a node in a Tree.
 type Node struct {
-	key           []byte
-	value         []byte
-	hash          []byte
-	nodeKey       *NodeKey
-	leftNodeKey   []byte
+	key     []byte
+	value   []byte
+	hash    []byte
+	nodeKey *NodeKey
+	// Legacy: LeftNodeHash
+	// v1: Left node ptr via Version/key
+	leftNodeKey []byte
+	// Legacy: RightNodeHash
+	// v1: Right node ptr via Version/key
 	rightNodeKey  []byte
 	size          int64
 	leftNode      *Node
 	rightNode     *Node
 	subtreeHeight int8
+	isLegacy      bool
 }
 
 var _ cache.Node = (*Node)(nil)
@@ -83,6 +88,9 @@ func NewNode(key []byte, value []byte) *Node {
 
 // GetKey returns the key of the node.
 func (node *Node) GetKey() []byte {
+	if node.isLegacy {
+		return node.hash
+	}
 	return node.nodeKey.GetKey()
 }
 
@@ -235,6 +243,7 @@ func MakeLegacyNode(hash, buf []byte) (*Node, error) {
 		nodeKey:       &NodeKey{version: ver},
 		key:           key,
 		hash:          hash,
+		isLegacy:      true,
 	}
 
 	// Read node body.
@@ -276,11 +285,11 @@ func (node *Node) String() string {
 	if node.rightNode != nil && node.rightNode.nodeKey != nil {
 		child += fmt.Sprintf("{right %v}", node.rightNode.nodeKey)
 	}
-	return fmt.Sprintf("Node{%s:%s@ %v:%v-%v %d-%d}#%s\n",
+	return fmt.Sprintf("Node{%s:%s@ %v:%x-%x %d-%d %x}#%s\n",
 		color.ColoredBytes(node.key, color.Green, color.Blue),
 		color.ColoredBytes(node.value, color.Cyan, color.Blue),
 		node.nodeKey, node.leftNodeKey, node.rightNodeKey,
-		node.size, node.subtreeHeight, child)
+		node.size, node.subtreeHeight, node.hash, child)
 }
 
 // clone creates a shallow copy of a node with its hash set to nil.
@@ -517,7 +526,7 @@ func (node *Node) writeHashBytes(w io.Writer, version int64) error {
 		// (e.g. ProofLeafNode.ValueHash)
 		valueHash := sha256.Sum256(node.value)
 
-		err = encoding.EncodeBytes(w, valueHash[:])
+		err = encoding.Encode32BytesHash(w, valueHash[:])
 		if err != nil {
 			return fmt.Errorf("writing value, %w", err)
 		}
@@ -525,11 +534,11 @@ func (node *Node) writeHashBytes(w io.Writer, version int64) error {
 		if node.leftNode == nil || node.rightNode == nil {
 			return ErrEmptyChild
 		}
-		err = encoding.EncodeBytes(w, node.leftNode.hash)
+		err = encoding.Encode32BytesHash(w, node.leftNode.hash)
 		if err != nil {
 			return fmt.Errorf("writing left hash, %w", err)
 		}
-		err = encoding.EncodeBytes(w, node.rightNode.hash)
+		err = encoding.Encode32BytesHash(w, node.rightNode.hash)
 		if err != nil {
 			return fmt.Errorf("writing right hash, %w", err)
 		}
@@ -595,7 +604,7 @@ func (node *Node) writeBytes(w io.Writer) error {
 			return fmt.Errorf("writing value, %w", err)
 		}
 	} else {
-		err = encoding.EncodeBytes(w, node.hash)
+		err = encoding.Encode32BytesHash(w, node.hash)
 		if err != nil {
 			return fmt.Errorf("writing hash, %w", err)
 		}
@@ -615,7 +624,7 @@ func (node *Node) writeBytes(w io.Writer) error {
 			return fmt.Errorf("writing mode, %w", err)
 		}
 		if mode&ModeLegacyLeftNode != 0 { // legacy leftNodeKey
-			err = encoding.EncodeBytes(w, node.leftNodeKey)
+			err = encoding.Encode32BytesHash(w, node.leftNodeKey)
 			if err != nil {
 				return fmt.Errorf("writing the legacy left node key, %w", err)
 			}
@@ -634,7 +643,7 @@ func (node *Node) writeBytes(w io.Writer) error {
 			return ErrRightNodeKeyEmpty
 		}
 		if mode&ModeLegacyRightNode != 0 { // legacy rightNodeKey
-			err = encoding.EncodeBytes(w, node.rightNodeKey)
+			err = encoding.Encode32BytesHash(w, node.rightNodeKey)
 			if err != nil {
 				return fmt.Errorf("writing the legacy right node key, %w", err)
 			}